Wordlist pthread support
This commit adds pthread support for the Wordlist dictionary. The
strategy is: we generate up to 8 worker threads that wait for new
files in a queue that is modified by a thread that does the path
traversal and adds the files as soon as it finds one. The worker
threads, thus, choose a file and process it.

Unfortunately, this approach does not work as expected for Aspell,
which is not *totally* thread-safe and requires a very complicated
strategy to parallelize the work. I have my doubts about any
performance gain, if any, more info here:
https://lists.gnu.org/archive/html/aspell-user/2005-10/msg00011.html
(Thread Safety Question).