Enhancement: Check entire identifiers and splitting if necessary
Checking identifiers is a bit complicated: if we only separate
the words (as it was previously), we can identify things like
semaphore_open (since both words are perfectly valid), but we
would not be able to identify 'sem_open' for example, since
'sem' it is not valid.

In order to decrease the false positives rate, this commit
introduces a hybrid approach: it checks whether the complete
word exists, if it does not, divides it, and checks each word
individually. Thus, the chances of correctly classifying an
identifier increase considerably.