Bug Fix: Trying to join already terminated threads
In the current state, Sourcery waits for the worker threads to finish
working with a 'pthread_join', but when they interrupt the work they
'kill' themselves instead of just returning, as expected by the main,
thus, memory leaks are reported to Valgrind. This commit just removes
pthread_exit and returns them naturally, as it should be.