Enhancement: Reverse Check Support
The Linux checkpatch.pl script uses a slightly different spellchecking
approach than Sourcery: instead of having a list of words and issuing
errors for all that are not on the list, checkpatch has a list of
errors and issues errors for all words that are present in the list.

This small difference causes the number of false positives to
drastically decrease, since we are really checking for common typing
errors, instead of waiting for Sourcery to know all the (correct) words
used in the src tree.

Therefore, to take advantage of the facilities already present in
Sourcery, this commit adds this type of approach, which can be obtained
via `make CHECK=reverse`.

Please note that Sourcery uses the same 'reverse' word list as the Linux
kernel, which is GPLv2, so if this license is a problem for your use
case, _do not_ use this feature.