Enhancement: Use open_memstream instead of array_list
This commit removes the use of array list for temporary storage of
misspellings in favor of using open_memstream. Using a buffer/stream
provides us with a number of advantages:

1) Simpler code and fewer lines of code
2) Less memory usage
3) Better utilization of threads: when in the dump portion, the thread
spends less time in the critical section, since the output is already
done and does not need to be formatted.
4) Less amount of syscalls: since write()'s performed with the largest
buffer size possible, we lose less time between system calls.

That said, these changes bring a good performance increase in the code,
depending on the size of the output: by analyzing the Linux kernel with
-s and -a, you get a performance gain of 18 and 100% respectively.