Enhancement: New words in reverse word list
This commit adds 2046 new words to the Sourcery error list (reverse
wordlist). Thus, the number of misspellings in the Linux src tree
increases from 2820 to 3957. It is important to note that, in a
visual analysis, there does not seem to be a large amount of false
positives.

It is important to note that this list of words was obtained from a
list of common mistakes made by Wikipedia editors. Although not
directly related to IT, this word list considerably increases the
number of detections (by 40%), and therefore I consider it to be
quite valid.

Other wordlists are available here:
https://www.dcs.bbk.ac.uk/~ROGER/corpora.html, and may be
considered for inclusion in the future.